+++ MonoEngine/source/rhi/agnostic/IndexBuffer.hh
@@ -11,14 +11,14 @@ namespace Monoworks::RHI
     public:
         virtual ~IIndexBuffer() {};
 
-        virtual void SetData( void* pData, u64 size, u64 offset = 0 ) NOEXCEPT = 0;
+        virtual void SetData( void* pData, u32 size, u32 offset = 0 ) NOEXCEPT = 0;
 
         // TODO: Remove VkCommandBuffer on the agnostic side
         virtual void Upload( VkCommandBuffer* pCmdBuffer ) NOEXCEPT = 0;
 
-        NODISCARD virtual u64 GetCount() NOEXCEPT = 0;
+        NODISCARD virtual u32 GetCount() NOEXCEPT = 0;
 
-        NODISCARD static Ref<IIndexBuffer> Create( u64 size ) NOEXCEPT;
-        NODISCARD static Ref<IIndexBuffer> Create( void* pData, u64 size, u64 offset = 0, bool autoupload = false ) NOEXCEPT;
+        NODISCARD static Ref<IIndexBuffer> Create( u32 size ) NOEXCEPT;
+        NODISCARD static Ref<IIndexBuffer> Create( void* pData, u32 size, u32 offset = 0, bool autoupload = false ) NOEXCEPT;
     };
 }
